AI_Service/requirements.txt
flask
flask-cors
sentence-transformers
faiss-cpu
numpy
mysql-connector-python
groq
python-dotenv
